talentscout_hiring_assistant/requirements.txt
streamlit
openai
python-dotenv
tqdm
